feat: update Purchase and PurchaseLineItem to use BigDecimal for quantities and totals to support partial values

+++ src/main/java/com/spiceflow/backend/purchase/dto/request/PurchaseLineItemRequest.java
@@ -14,11 +14,11 @@ public record PurchaseLineItemRequest(
     
     @NotNull
     @PositiveOrZero
-    Integer noOfBoxes,
+    BigDecimal noOfBoxes,
     
     @NotNull
     @PositiveOrZero
-    Integer soldQuantity,
+    BigDecimal soldQuantity,
     
     String unitType,
